+++ src/memory/KernelMemoryAllocator.h
@@ -21,7 +21,7 @@ namespace memory {
  */
 namespace KernelMemoryAllocator {
 
-    void* malloc(uint64_t size);
+    void* malloc(size_t size);
     void free(void* addr);
 
     /**
